Explain AI spend and show outbound traffic on Operations
Every AI call carries a stage (a CV build: rubric, author, review,
review_retry) so a build's cost can be itemised, not only summed.
Server-side search charges are priced into recorded cost, a failed call
is priced at the model that served it, and cancelled sibling batches and
stalled streams are classified apart from real failures. Cost sums cast
to double precision before adding. ai_calls is pruned after 13 months
and the traffic rollup after 400 days.

Operations gains outbound traffic per host over seven days with a
week-over-week delta, 304 share, throttled and blocked counts and a
bucketed p95; AI performance per feature with p50/p95 latency, cache-hit
share and the failure split; cost per build by stage and cost per
scored role; and method, requests and revalidations on the largest scan
inputs. DEPLOY.md documents the ledgers, retention, what stays a log
line, and how to answer the four operational questions.

+++ packages/db/src/scaling.ts
@@ -26,7 +26,9 @@ export async function workloadMetrics(db: Db) {
       (select count(*)::int from companies c where c.status='active' and c.added_at < now()-interval '1 day'
         and not exists (select 1 from career_sources cs where cs.company_id=c.id and cs.last_ok_scan_at > now()-interval '1 day')) as overdue_companies,
       (select count(*)::int from discovery_sources where enabled=true and next_run_at < now()-interval '1 day') as overdue_discovery,
-      (select coalesce(sum(amount),0)::float from ai_reservations where expires_at > now()) as reserved_usd
+      -- amount is real, as cost_usd is: sum() over it accumulates in single precision, so a month
+      -- of small holds drifts. Widen each value before adding, not the total afterwards.
+      (select coalesce(sum(amount::float8),0) from ai_reservations where expires_at > now()) as reserved_usd
     from queue`);
   const row = result.rows[0]!;
   return {
